Move SensitivityValueKey in seperated file and add small test for coverage.
